:truck: chore(api): move certification jury done event
