+++ api/tests/unit/domain/events/handle-certification-rescoring_test.js
@@ -1,8 +1,8 @@
-import { CertificationJuryDone } from '../../../../lib/domain/events/CertificationJuryDone.js';
 import { _forTestOnly } from '../../../../lib/domain/events/index.js';
 import { ChallengeDeneutralized } from '../../../../src/certification/evaluation/domain/events/ChallengeDeneutralized.js';
 import { ChallengeNeutralized } from '../../../../src/certification/evaluation/domain/events/ChallengeNeutralized.js';
 import { CertificationCourseRejected } from '../../../../src/certification/session-management/domain/events/CertificationCourseRejected.js';
+import { CertificationJuryDone } from '../../../../src/certification/session-management/domain/events/CertificationJuryDone.js';
 import { CertificationAssessment } from '../../../../src/certification/session-management/domain/models/CertificationAssessment.js';
 import { AlgorithmEngineVersion } from '../../../../src/certification/shared/domain/models/AlgorithmEngineVersion.js';
 import { ABORT_REASONS } from '../../../../src/certification/shared/domain/models/CertificationCourse.js';
